pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codemaxxx"
version = "0.1.0"
description = "CodeMaxxx — AI coding agent for the terminal, powered by Ollama"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
  "rich>=13.7",
  "prompt-toolkit>=3.0",
  "httpx>=0.27",
  "psycopg2-binary>=2.9",
]

[project.scripts]
codemaxxx-tui = "codemaxxx.main:main"

[tool.setuptools.packages.find]
where = ["src"]
